+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/persistentlock/LockStore.java
@@ -15,7 +15,6 @@
  */
 package com.palantir.atlasdb.persistentlock;
 
-import java.nio.charset.StandardCharsets;
 import java.util.List;
 import java.util.Set;
 import java.util.UUID;
@@ -25,6 +24,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.common.annotations.VisibleForTesting;
+import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
 import com.palantir.atlasdb.AtlasDbConstants;
 import com.palantir.atlasdb.keyvalue.api.CheckAndSetException;
@@ -35,7 +35,7 @@ import com.palantir.atlasdb.keyvalue.api.RowResult;
 import com.palantir.atlasdb.keyvalue.api.Value;
 
 /**
- * LockStore manages {@link LockEntry} objects, specifically for the "Deletion Lock" (to be taken out by backup and
+ * LockStore manages {@link LockEntry} objects, specifically for the "Backup Lock" (to be taken out by backup and
  * sweep).
  *
  * The PERSISTED_LOCKS_TABLE contains exactly one element, either LOCK_OPEN or (lock taken for some REASON), giving
@@ -43,8 +43,8 @@ import com.palantir.atlasdb.keyvalue.api.Value;
  *
  * "taken for SWEEP (uuid1)" <- - - - -> OPEN < - - - -> "taken for BACKUP (uuid2)"
  *
- * acquireLock(REASON) attempts to move the lock state from "OPEN" to "taken because REASON", returning a LockEntry
- * that holds a unique identifier for that occasion of taking the lock.
+ * acquireBackupLock(REASON) attempts to move the lock state from "OPEN" to "taken because REASON",
+ * returning a LockEntry that holds a unique identifier for that occasion of taking the lock.
  *
  * releaseLock(LockEntry) attempts to move the state from "taken because REASON" to "OPEN", but only succeeds if the
  * LockEntry currently stored matches the one passed in, ensuring that nobody released the lock when we weren't looking.
@@ -53,15 +53,16 @@ import com.palantir.atlasdb.keyvalue.api.Value;
  * If we fail to do this, it's because someone else also created the table, and populated it before we did.
  * This is actually OK - all we care about is that we're in the state machine _somewhere_.
  */
-public final class LockStore {
+@SuppressWarnings("checkstyle:FinalClass") // Non-final as we'd like to mock it.
+public class LockStore {
     private static final Logger log = LoggerFactory.getLogger(LockStore.class);
 
-    private static final String ROW_NAME = "DeletionLock";
+    private static final String BACKUP_LOCK_NAME = "BackupLock";
     private final KeyValueService keyValueService;
 
     public static final LockEntry LOCK_OPEN = ImmutableLockEntry.builder()
-            .rowName(ROW_NAME)
-            .lockId("-1")
+            .lockName(BACKUP_LOCK_NAME)
+            .instanceId(UUID.fromString("0-0-0-0-0"))
             .reason("Available")
             .build();
 
@@ -78,14 +79,15 @@ public final class LockStore {
         return lockStore;
     }
 
-    public LockEntry acquireLock(String reason) throws CheckAndSetException {
-        LockEntry lockEntry = generateUniqueLockEntry(reason);
+    public LockEntry acquireBackupLock(String reason) throws CheckAndSetException {
+        LockEntry lockEntry = generateUniqueBackupLockEntry(reason);
         CheckAndSetRequest request = CheckAndSetRequest.singleCell(AtlasDbConstants.PERSISTED_LOCKS_TABLE,
                 lockEntry.cell(),
                 LOCK_OPEN.value(),
                 lockEntry.value());
 
         keyValueService.checkAndSet(request);
+        log.info("Successfully acquired the persistent lock: {}", lockEntry);
         return lockEntry;
     }
 
@@ -96,6 +98,7 @@ public final class LockStore {
                 LOCK_OPEN.value());
 
         keyValueService.checkAndSet(request);
+        log.info("Successfully released the persistent lock.");
     }
 
     @VisibleForTesting
@@ -108,9 +111,13 @@ public final class LockStore {
         return results.stream().map(LockEntry::fromRowResult).collect(Collectors.toSet());
     }
 
-    private LockEntry generateUniqueLockEntry(String reason) {
-        String randomLockId = UUID.randomUUID().toString();
-        return ImmutableLockEntry.builder().rowName(ROW_NAME).lockId(randomLockId).reason(reason).build();
+    private LockEntry generateUniqueBackupLockEntry(String reason) {
+        UUID randomLockId = UUID.randomUUID();
+        return ImmutableLockEntry.builder()
+                .lockName(BACKUP_LOCK_NAME)
+                .instanceId(randomLockId)
+                .reason(reason)
+                .build();
     }
 
     static class LockStorePopulator {
@@ -131,12 +138,20 @@ public final class LockStore {
                 // This can happen if multiple LockStores are started at once. We don't actually mind.
                 // All we care about is that we're in the state machine of "LOCK_OPEN"/"LOCK_TAKEN".
                 // It still might be interesting, so we'll log it.
-                List<String> values = e.getActualValues().stream()
-                        .map(v -> new String(v, StandardCharsets.UTF_8))
-                        .collect(Collectors.toList());
-                log.info("Encountered a CheckAndSetException when creating the LockStore. This means that two "
+                List<String> values = getActualValues(e);
+                log.debug("Encountered a CheckAndSetException when creating the LockStore. This means that two "
                         + "LockStore objects were created near-simultaneously, and is probably not a problem. "
                         + "For the record, we observed these values: {}", values);
+            }
+        }
+
+        private List<String> getActualValues(CheckAndSetException ex) {
+            if (ex.getActualValues() != null) {
+                return ex.getActualValues().stream()
+                        .map(v -> LockEntry.fromStoredValue(v).toString())
+                        .collect(Collectors.toList());
+            } else {
+                return ImmutableList.of();
             }
         }
     }